requirements.txt
aiogram==3.5.0
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
pytz==2024.1
apscheduler==3.10.4
openpyxl==3.1.2
psutil

